it working in local

+++ experiments/explanation_generation/ExplanationGeneration_food.py
@@ -56,8 +56,8 @@ def main():
     num_ftrs = model.fc.in_features
     model.fc = nn.Linear(num_ftrs, num_of_classes)
     model = model.to(device)
-    model.load_state_dict(torch.load(model_path, map_location=device )["model_state"]).to(dtype=torch.half)
-
+    model.load_state_dict(torch.load(model_path, map_location=device )["model_state"])
+    model = model.type(torch.HalfTensor)
     transform_train = transforms.Compose([transforms.Resize(image_size), # transforms.RandomHorizontalFlip(),
                                     transforms.ToTensor(),
                                     transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])])
